+++ src/Rule/IdRule.php
@@ -3,14 +3,15 @@ namespace Pyncer\Validation\Rule;
 
 use Pyncer\Data\Mapper\MapperInterface;
 use Pyncer\Data\MapperQuery\MapperQueryInterface;
-use Pyncer\Validation\Rule\IntRule;
+use Pyncer\Validation\Rule\AbstractRule;
+use Stringable;
 
 use function array_key_exists;
 use function intval;
 use function is_string;
 use function trim;
 
-class IdRule extends IntRule
+class IdRule extends AbstractRule
 {
     /**
      * @var array<int, bool>
@@ -22,49 +23,27 @@ class IdRule extends IntRule
      * @param string $column The name of the database column that stores the id.
      * @param \Pyncer\Data\MapperQuery\MapperQueryInterface $mapperQuery
      *      Optional mapper query to limit query results.
-     * @param bool $allowNull When true, null vlaues are valid.
-     * @param bool $allowEmpty When true, empty values are valid.
      */
     public function __construct(
         private MapperInterface $mapper,
         private string $column = 'id',
         private ?MapperQueryInterface $mapperQuery = null,
-        bool $allowNull = false,
-        bool $allowEmpty = false,
     ) {
         parent::__construct(
-            minValue: 0,
-            allowNull: $allowNull,
-            allowEmpty: $allowEmpty,
+            allowNull: true,
+            allowEmpty: true,
+            empty: 0,
         );
     }
 
     /**
      * @inheritdoc
      */
-    public function isValid(mixed $value): bool
+    protected function isValidValue(mixed $value): bool
     {
-        if (!parent::isValid($value)) {
+        if (!is_int($value)) {
             return false;
         }
-
-        if (!is_int($value) || $value === $this->empty) {
-            return true;
-        }
-
-        return $this->isValidId($value);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function isValidId(mixed $value): bool
-    {
-        if (is_string($value)) {
-            $value = trim($value);
-        }
-
-        $value = intval($value);
 
         if ($value <= 0) {
             return false;
@@ -88,31 +67,5 @@ class IdRule extends IntRule
         }
 
         return $this->tests[$value];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function clean(mixed $value): mixed
-    {
-        $value = parent::clean($value);
-
-        if (!is_int($value) || $value === $this->empty) {
-            return $value;
-        }
-
-        if (!$this->isValidId($value)) {
-            if ($this->allowNull) {
-                return null;
-            }
-
-            if ($this->allowEmpty) {
-                return $this->empty;
-            }
-
-            return '';
-        }
-
-        return $value;
     }
 }